Revert "feat: add serverFn function"
This reverts commit 272b04fbf3ee2e7e06548f892337895b2a90fa0f.

+++ apps/www/src/components/auth/profile/delete-dialog.tsx
@@ -1,7 +1,7 @@
 import { db } from '@/core/database';
 import { users } from '@/core/database/schema';
 import { useAuth } from '@/www/contexts/auth';
-import { serverFnRequiresAuth } from '@/www/server/server-fn';
+import { auth } from '@/www/server/auth';
 import { useNavigate } from '@solidjs/router';
 import { createMutation } from '@tanstack/solid-query';
 import { eq } from 'drizzle-orm';
@@ -17,10 +17,17 @@ import {
   DialogTrigger,
 } from '../../ui/dialog';
 
-const deleteUser = serverFnRequiresAuth(async ({ user }) => {
+async function deleteUser() {
+  'use server';
+  const { user } = auth();
+  if (!user) {
+    throw new Error('Unauthorized');
+  }
+
   await db.delete(users).where(eq(users.id, user.id));
+
   return { success: true };
-});
+}
 
 export const DeleteProfileDialog = () => {
   const navigate = useNavigate();
