+++ apps/www/src/components/chat/sidebar/delete-chat-button.tsx
@@ -12,17 +12,24 @@ import {
 } from '@/www/components/ui/dialog';
 import { Tooltip, TooltipContent, TooltipTrigger } from '@/www/components/ui/tooltip';
 import { useChatStore } from '@/www/contexts/chat';
-import { serverFnRequiresAuth } from '@/www/server/server-fn';
+import { auth } from '@/www/server/auth';
 import type { DialogTriggerProps } from '@kobalte/core/dialog';
 import { useNavigate } from '@solidjs/router';
 import { createMutation, useQueryClient } from '@tanstack/solid-query';
 import { and, eq } from 'drizzle-orm';
 import { Trash } from 'lucide-solid';
 
-const deleteChat = serverFnRequiresAuth(async ({ user }, chatId: string) => {
+const deleteChat = async (chatId: string) => {
+  'use server';
+  const { user } = auth();
+  if (!user) {
+    throw new Error('User is not authenticated');
+  }
+
   await db.delete(chats).where(and(eq(chats.userId, user.id), eq(chats.id, chatId)));
+
   return { success: true };
-});
+};
 
 export type DeleteChatButtonProps = {
   chat: Chat;
